+++ client/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './src/App.jsx';
+import Review from './src/Review.jsx';
 
-ReactDOM.render(<App />, document.getElementById('reviews'));
+ReactDOM.render(<Review />, document.getElementById('review'));
