+++ client/src/Review.jsx
@@ -4,10 +4,10 @@ import Rating from './Rating.jsx';
 import SearchBar from './SearchBar.jsx';
 import Comment from './Comment.jsx';
 import PageNumber from './PageNumber.jsx';
-import { AppContainer, Top } from '../appStyles.jsx';
+import { AppContainer, Top } from '../reviewStyles.jsx';
 
 
-class App extends React.Component {
+class Review extends React.Component {
   constructor(props) {
     super(props);
 
@@ -107,4 +107,4 @@ class App extends React.Component {
   }
 }
 
-export default App;
+export default Review;
