+++ public/index.html
@@ -3,7 +3,7 @@
   <head>
   </head>
   <body>
-    <div id="reviews"></div>
+    <div id="review"></div>
     <script src="./dist/bundle.js"></script>
   </body>
 </html>
